// src/config/env.d.ts
import "dotenv/config";
export declare const env: {
    nodeEnv: string;
    port: number;
    frontendOrigin: string;
    smtpHost: string | undefined;
    smtpPort: number;
    smtpSecure: boolean;
    smtpUser: string | undefined;
    smtpPass: string | undefined;
    mailFrom: string | undefined;
    mailTo: string | undefined;
};
//# sourceMappingURL=env.d.ts.map